Refactored apointmentUI to inherit from a base class
